refactor mobile detect
